Add files via upload
The app.rs file is meant to serve as a guide for building apps for the device, I do need to rework it since I previously made it thinking we'd use lvgl. The input_handler strictly tests if it can read from a keyboard or hardware, and the main file is the previous c++ code of the main menu now in rust.